moved code from component to service
